fix: Organize all images in one images folder
