+++ src/components/Salut.jsx
@@ -1,19 +1,16 @@
-import React from 'react'
-import phoneApp  from "../assets/phone-App.png";
-import line from "../assets/line.png"
-import "../styles/Salut.css"
-
-
+import React from "react";
+import phoneApp from "../assets/images/phone-App.png";
+import line from "../assets/images/line.png";
+import "../styles/Salut.css";
 
 const Salut = () => {
-  return ( 
-
+  return (
     <div>
-    <img src={phoneApp} alt="phoneApp "/>
-    <div className="string">
-    <img src= {line} alt="line" />
+      <img src={phoneApp} alt="phoneApp " />
+      <div className="string">
+        <img src={line} alt="line" />
+      </div>
     </div>
-    </div>  
-);
+  );
 };
 export default Salut;
